# server/src/db/schema.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export interface RefreshToken {
  createdAt: Generated<Date | null>;
  expiresAt: Date;
  id: string;
  isRevoked: Generated<number | null>;
  tokenHash: string;
  updatedAt: Generated<Date | null>;
  userId: string;
}

export interface User {
  createdAt: Generated<Date | null>;
  email: string;
  googleSub: string | null;
  id: string;
  isActive: Generated<number | null>;
  isEmailVerified: Generated<number | null>;
  name: string;
  password: string | null;
  updatedAt: Generated<Date | null>;
}

export interface DB {
  refresh_tokens: RefreshToken;
  users: User;
}